requirements.txt
discord.py
discord.py[voice]
colorlog
PyNaCl
dnspython
async-timeout
peewee
python-dotenv